Close other editors found on same tabbar
